+++ lib/models/unused_asset_model.dart
@@ -1,4 +1,3 @@
-
 class UnusedAssetModel {
   String fileName;
   String filePath;