+++ lib/services/config_services.dart
@@ -7,13 +7,15 @@ class ConfigServices {
 
   Future<bool> create() async {
     try {
-      String templateUrl = "https://raw.githubusercontent.com/yusriltakeuchi/assets_cleaner/master/template/assets_cleaner.yaml";
+      String templateUrl =
+          "https://raw.githubusercontent.com/yusriltakeuchi/assets_cleaner/master/template/assets_cleaner.yaml";
+
       /// Downloading from the url and save to [path]
       final request = await HttpClient().getUrl(Uri.parse(templateUrl));
       final response = await request.close();
       response.pipe(File('assets_cleaner.yaml').openWrite());
       return true;
-    } catch(e) {
+    } catch (e) {
       return false;
     }
   }