+++ lib/utils/code_utils.dart
@@ -15,7 +15,7 @@ class CodeUtils {
       // Only add in List if file in path is supported
       if (regExp.hasMatch('$dir')) {
         listFiles.add(dir);
-      }   
+      }
     });
     return listFiles;
   }
@@ -24,15 +24,14 @@ class CodeUtils {
     String currentPath = FileUtils.instance.getCurrentPath;
     String libDirectory = "$currentPath/lib";
     final libCodes = await _scanCodes(Directory(libDirectory));
-    
+
     for (var code in libCodes) {
       final path = code.path.replaceAll(currentPath, "");
       final content = File("$currentPath$path").readAsStringSync();
       if (content.contains(asset) == true) {
         return true;
-      } 
+      }
     }
     return false;
   }
-
 }